Resolve merge conflict in .gitmodules
